+++ FE/src/hooks/query/useMemberQuery.ts
@@ -40,6 +40,7 @@ export const useGetProgramMembersByActive = ({
   return useQuery({
     queryKey: [API.MEMBER.ACTIVE_STATUS(programId), status],
     queryFn: () => getProgramMembersByActiveStatus(programId, status),
+    staleTime: 1000 * 60 * 5,
   });
 };
 
@@ -50,6 +51,7 @@ export const useGetProgramMembersByAttend = ({
   return useQuery({
     queryKey: [API.MEMBER.ATTEND_STATUS(programId), status],
     queryFn: () => getProgramMembersByAttendStatus(programId, status),
+    staleTime: 1000 * 60 * 5,
   });
 };
