+++ FE/src/hooks/query/useUserQuery.ts
@@ -33,6 +33,7 @@ export const useGetMyAttendStatus = (programId: number) => {
   return useQuery({
     queryKey: [API.USER.ATTEND_STATUS(programId)],
     queryFn: () => getMyAttendStatus(programId),
+    staleTime: 1000 * 60 * 5,
   });
 };
 
@@ -41,6 +42,10 @@ interface PutMyAttendStatus {
   beforeAttendStatus: AttendStatus;
 }
 
+/**
+ * 이전에 직접 참석 여부를 변경할 떄 사용하던 api 입니다.
+ * 현재는 사용하지 않습니다.
+ */
 export const usePutMyAttendStatus = ({
   programId,
   beforeAttendStatus,
